Allow for compilation without obs frontend
Using the cmake ENABLE_UI switch (already defined in OBS)
