+++ rtsp_frontend.h
@@ -1,0 +1,9 @@
+#ifdef OBS_RTSPSERVER_ENABLE_UI
+#ifndef RTSP_FRONTEND_H
+#define RTSP_FRONTEND_H
+
+void rtsp_frontend_load();
+void rtsp_frontend_unload();
+
+#endif // RTSP_FRONTEND_H
+#endif // OBS_RTSPSERVER_ENABLE_UI
